# api/openapi.yaml
openapi: 3.1.0
info:
  title: QianuHub 闲鱼助手 API
  version: 1.0.0
  description: 版本化 HTTP API 的唯一契约源。
servers:
  - url: /
paths:
...
  "/api/v1/chat/images":
    post:
      operationId: postApiV1ChatImages
      summary: postApiV1ChatImages
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [account_id, chat_id, peer_user_id, image]
              properties:
                account_id: {type: string}
                chat_id: {type: string}
                peer_user_id: {type: string}
                peer_name: {type: string}
                peer_avatar_url: {type: string}
                item_id: {type: string}
                item_title: {type: string}
                image: {type: string, format: binary}
      responses:
        '201':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessageEnvelope'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 消息已发送但本地状态收口失败，或其他统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatSendErrorResponse'
        '502':
          description: 平台发送失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatSendErrorResponse'
...
  "/api/v1/chat/messages":
    get:
      operationId: getApiV1ChatMessages
      summary: getApiV1ChatMessages
      security: [{cookieAuth: []}]
      parameters:
        - name: account_id
          in: query
          required: true
          schema: {type: string}
        - name: chat_id
          in: query
          required: true
          schema: {type: string}
        - name: cursor
          in: query
          required: false
          schema: {type: integer}
        - name: before_id
          in: query
          required: false
          schema: {type: integer}
        - name: limit
          in: query
          required: false
          schema: {type: integer, minimum: 1, maximum: 200}
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessagePage'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: postApiV1ChatMessages
      summary: postApiV1ChatMessages
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatMessageRequest'
      responses:
        '201':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessageEnvelope'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 消息已发送但本地状态收口失败，或其他统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatSendErrorResponse'
        '502':
          description: 平台发送失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatSendErrorResponse'
  "/api/v1/chat/read":
    post:
      operationId: postApiV1ChatRead
      summary: postApiV1ChatRead
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatReadRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  "/api/v1/chat/sessions":
    delete:
      operationId: deleteApiV1ChatSessions
      summary: deleteApiV1ChatSessions
      security: [{cookieAuth: []}]
      parameters:
        - name: account_id
          in: query
          required: true
          schema: {type: string}
        - name: chat_id
          in: query
          required: true
          schema: {type: string}
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      operationId: getApiV1ChatSessions
      summary: getApiV1ChatSessions
      security: [{cookieAuth: []}]
      parameters:
        - name: account_id
          in: query
          required: true
          schema: {type: string}
        - name: cursor
          in: query
          required: false
          schema: {type: integer}
        - name: stored_cursor
          in: query
          required: false
          schema: {type: string}
        - name: refresh
          in: query
          required: false
          schema: {type: integer, enum: [0, 1]}
        - name: limit
          in: query
          required: false
          schema: {type: integer, minimum: 1, maximum: 200}
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatSessionPage'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: 平台联系人刷新失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
...
    ChatSessionPage:
      type: object
      required: [sessions, has_more, platform_has_more, stored_has_more]
      properties:
        sessions: {type: array, items: {$ref: '#/components/schemas/ChatSession'}}
        has_more: {type: boolean}
        next_cursor: {type: integer}
        next_stored_cursor: {type: string}
        platform_has_more: {type: boolean}
        stored_has_more: {type: boolean}
    ChatMessagePage:
      type: object
      required: [messages, has_more]
      properties:
        messages: {type: array, items: {$ref: '#/components/schemas/ChatMessage'}}
        has_more: {type: boolean}
        next_cursor: {type: integer}
        session: {$ref: '#/components/schemas/ChatSession'}
    ChatMessageEnvelope:
      type: object
      required: [message]
      properties:
        message: {$ref: '#/components/schemas/ChatMessage'}
    ChatSendErrorResponse:
      type: object
      required: [code, message]
      properties:
        code: {type: string}
        message: {type: string}
        request_id: {type: string}
        details:
          type: object
          properties:
            outgoing_message: {$ref: '#/components/schemas/ChatMessage'}
      additionalProperties: true
...
    ChatMessageRequest:
      type: object
      required: [account_id, chat_id, peer_user_id, text]
      properties:
        account_id: {type: string}
        chat_id: {type: string}
        peer_user_id: {type: string}
        peer_name: {type: string}
        item_id: {type: string}
        item_title: {type: string}
        text: {type: string, maxLength: 2000}
...
    ChatReadRequest:
      type: object
      required: [account_id, chat_id]
      properties:
        account_id: {type: string}
        chat_id: {type: string}
        message_ids: {type: array, items: {type: object, additionalProperties: {oneOf: [{type: string}, {type: number}]}}}
...
    OperationResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string
        requires_relogin:
          type: boolean
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
        details:
          type: object
      additionalProperties: true